+++ spec/models/auto_detection_spec.rb
@@ -37,6 +37,15 @@ RSpec.describe 'auto-detection' do
     expect { LegacyUser.find(UNKNOWN_UUID) }.to raise_error(ActiveRecord::RecordNotFound)
   end
 
+  it 'raises when translating on a model without polyid' do
+    LegacyUser.create!(name: 'Legacy')
+
+    expect { LegacyUser.id_for(UNKNOWN_UUID) }.to raise_error(/not configured with polyid/)
+    expect { LegacyUser.uuid_for(1) }.to raise_error(/not configured with polyid/)
+    expect { LegacyUser.ids_for([UNKNOWN_UUID]) }.to raise_error(/not configured with polyid/)
+    expect { LegacyUser.uuids_for([1]) }.to raise_error(/not configured with polyid/)
+  end
+
   it 'can disable auto-detection globally' do
     PolyId.auto_detect = false
     user = User.create!(name: 'User', uuid: UNKNOWN_UUID)